+++ ktlint-ruleset-standard/src/test/kotlin/com/pinterest/ktlint/ruleset/standard/rules/DiscouragedCommentLocationRuleTest.kt
@@ -501,15 +501,21 @@ class DiscouragedCommentLocationRuleTest {
                 """
                 class Foo1(
                     // some comment
-                    val bar: Bar
+                    val bar: Bar,
+                    // some comment
+                    val bar: Bar,
                 )
                 class Foo2(
                     /* some comment */
-                    val bar: Bar
+                    val bar: Bar,
+                    /* some comment */
+                    val bar: Bar,
                 )
                 class Foo3(
                     /** some comment */
-                    val bar: Bar
+                    val bar: Bar,
+                    /** some comment */
+                    val bar: Bar,
                 )
                 """.trimIndent()
             discouragedCommentLocationRuleAssertThat(code).hasNoLintViolations()
@@ -531,11 +537,11 @@ class DiscouragedCommentLocationRuleTest {
                     val bar: /** some comment */ Bar
                 )
                 """.trimIndent()
-            @Suppress("ktlint:standard:parameter-list-wrapping", "ktlint:standard:max-line-length")
+            @Suppress("ktlint:standard:argument-list-wrapping", "ktlint:standard:max-line-length")
             discouragedCommentLocationRuleAssertThat(code)
                 .hasLintViolationsWithoutAutoCorrect(
-                    LintViolation(3, 9, "A comment in a 'value_parameter' is only allowed when placed on a new line before this element"),
-                    LintViolation(7, 14, "A comment in a 'value_parameter' is only allowed when placed on a new line before this element"),
+                    LintViolation(3, 9, "A (block or EOL) comment inside or on same line after a 'value_parameter' is not allowed. It may be placed on a separate line above."),
+                    LintViolation(7, 14, "A (block or EOL) comment inside or on same line after a 'value_parameter' is not allowed. It may be placed on a separate line above."),
                     LintViolation(10, 14, "A kdoc in a 'value_parameter' is only allowed when placed on a new line before this element"),
                 )
         }
@@ -554,12 +560,12 @@ class DiscouragedCommentLocationRuleTest {
                     val bar: Bar /* some comment */
                 )
                 """.trimIndent()
-            @Suppress("ktlint:standard:parameter-list-wrapping", "ktlint:standard:max-line-length")
+            @Suppress("ktlint:standard:argument-list-wrapping", "ktlint:standard:max-line-length")
             discouragedCommentLocationRuleAssertThat(code)
                 .hasLintViolationsWithoutAutoCorrect(
-                    LintViolation(2, 18, "A comment in a 'value_parameter' is only allowed when placed on a new line before this element"),
-                    LintViolation(5, 18, "A comment in a 'value_parameter' is only allowed when placed on a new line before this element"),
-                    LintViolation(8, 18, "A comment in a 'value_parameter' is only allowed when placed on a new line before this element"),
+                    LintViolation(2, 18, "A (block or EOL) comment inside or on same line after a 'value_parameter' is not allowed. It may be placed on a separate line above."),
+                    LintViolation(5, 18, "A (block or EOL) comment inside or on same line after a 'value_parameter' is not allowed. It may be placed on a separate line above."),
+                    LintViolation(8, 18, "A (block or EOL) comment inside or on same line after a 'value_parameter' is not allowed. It may be placed on a separate line above."),
                 )
         }
 
@@ -574,7 +580,6 @@ class DiscouragedCommentLocationRuleTest {
                     val bar: Bar, /* some comment */
                 )
                 """.trimIndent()
-            @Suppress("ktlint:standard:parameter-list-wrapping", "ktlint:standard:max-line-length")
             discouragedCommentLocationRuleAssertThat(code)
                 .hasLintViolationsWithoutAutoCorrect(
                     LintViolation(2, 19, "A comment in a 'value_parameter_list' is only allowed when placed on a separate line"),
@@ -636,7 +641,6 @@ class DiscouragedCommentLocationRuleTest {
                     >
                 class Foo2<Bar /* some comment */ >
                 """.trimIndent()
-            @Suppress("ktlint:standard:parameter-list-wrapping", "ktlint:standard:max-line-length")
             discouragedCommentLocationRuleAssertThat(code)
                 .hasLintViolationsWithoutAutoCorrect(
                     LintViolation(2, 9, "A comment in a 'type_parameter_list' is only allowed when placed on a separate line"),
@@ -653,7 +657,6 @@ class DiscouragedCommentLocationRuleTest {
                     Bar>
                 class FooBar2<Foo, /* some comment */ Bar>
                 """.trimIndent()
-            @Suppress("ktlint:standard:parameter-list-wrapping", "ktlint:standard:max-line-length")
             discouragedCommentLocationRuleAssertThat(code)
                 .hasLintViolationsWithoutAutoCorrect(
                     LintViolation(2, 10, "A comment in a 'type_parameter_list' is only allowed when placed on a separate line"),
@@ -715,7 +718,6 @@ class DiscouragedCommentLocationRuleTest {
                 val fooBar2: FooBar<Foo, /* some comment */
                     Bar>
                 """.trimIndent()
-            @Suppress("ktlint:standard:parameter-list-wrapping", "ktlint:standard:max-line-length")
             discouragedCommentLocationRuleAssertThat(code)
                 .hasLintViolationsWithoutAutoCorrect(
                     LintViolation(1, 26, "A comment in a 'type_argument_list' is only allowed when placed on a separate line"),
@@ -819,5 +821,32 @@ class DiscouragedCommentLocationRuleTest {
                     LintViolation(11, 27, "No comment expected at this location"),
                 )
         }
+    }
+
+    @Test
+    fun `Given value argument preceded by a KDOC plus an EOL or BLOCK comments on separate lines above`() {
+        val code =
+            """
+            class Foo(
+                /** Some kdoc */
+                /* Some comment */
+                val bar1: Int,
+                /** Some kdoc */
+                // Some comment
+                val bar2: Int,
+                /* Some comment */
+                /** Some kdoc */
+                val bar3: Int,
+                // Some comment
+                /** Some kdoc */
+                val bar4: Int,
+            )
+            """.trimIndent()
+        @Suppress("ktlint:standard:argument-list-wrapping", "ktlint:standard:max-line-length")
+        discouragedCommentLocationRuleAssertThat(code)
+            .hasLintViolationsWithoutAutoCorrect(
+                LintViolation(3, 5, "A (block or EOL) comment inside or on same line after a 'value_parameter' is not allowed. It may be placed on a separate line above."),
+                LintViolation(6, 5, "A (block or EOL) comment inside or on same line after a 'value_parameter' is not allowed. It may be placed on a separate line above."),
+            )
     }
 }